requirements.txt
Flask_Limiter>=1.0.1
Flask>=1.1.1
flask_restx>=0.5.0
Flask_Cors>=3.0.8
setuptools>=41.1.0
requests>=2.23.0
pandas>=1.3.0
numpy>=1.18.3
Werkzeug>=1.0.1
pillow==9.2.0
tensorflow==2.11.0
keras==2.11.0